refactor: db 관련 파일들 패키지 이동
