+++ app/src/main/java/woowacourse/omok/database/OmokContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok
+package woowacourse.omok.database
 
 import android.provider.BaseColumns
