+++ app/src/main/java/woowacourse/omok/database/OmokDBHelper.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok
+package woowacourse.omok.database
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
